fix(mobile): fixed InitZCNSDK for mobile building
