+++ zcncore/wallet_base.go
@@ -335,7 +335,6 @@ func CreateWallet(statusCb WalletCallback) error {
 	return nil
 }
 
-
 // registerToMiners can be used to register the wallet.
 func registerToMiners(wallet *zcncrypto.Wallet, statusCb WalletCallback) error {
 	result := make(chan *util.PostResponse)
